Reorganized client use cases
