+++ LBH.AdultSocialCare.Api/V1/UseCase/ClientsUseCases/Concrete/DeleteClientsUseCase.cs
@@ -1,9 +1,9 @@
-using LBH.AdultSocialCare.Api.V1.Gateways.Interfaces;
-using LBH.AdultSocialCare.Api.V1.UseCase.Interfaces;
 using System;
 using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Gateways.Interfaces;
+using LBH.AdultSocialCare.Api.V1.UseCase.ClientsUseCases.Interfaces;
 
-namespace LBH.AdultSocialCare.Api.V1.UseCase.ClientsUseCases
+namespace LBH.AdultSocialCare.Api.V1.UseCase.ClientsUseCases.Concrete
 {
     public class DeleteClientsUseCase : IDeleteClientsUseCase
     {
