+++ LBH.AdultSocialCare.Api/V1/UseCase/ClientsUseCases/Interfaces/IDeleteClientsUseCase.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Threading.Tasks;
 
-namespace LBH.AdultSocialCare.Api.V1.UseCase.Interfaces
+namespace LBH.AdultSocialCare.Api.V1.UseCase.ClientsUseCases.Interfaces
 {
     public interface IDeleteClientsUseCase
     {
