+++ LBH.AdultSocialCare.Api/V1/UseCase/ClientsUseCases/Interfaces/IGetClientsUseCase.cs
@@ -1,8 +1,8 @@
-using LBH.AdultSocialCare.Api.V1.Domain.ClientDomains;
 using System;
 using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Domain.ClientDomains;
 
-namespace LBH.AdultSocialCare.Api.V1.UseCase.Interfaces
+namespace LBH.AdultSocialCare.Api.V1.UseCase.ClientsUseCases.Interfaces
 {
     public interface IGetClientsUseCase
     {
