+++ LBH.AdultSocialCare.Api/V1/UseCase/ClientsUseCases/Interfaces/IUpsertClientsUseCase.cs
@@ -1,7 +1,7 @@
-using LBH.AdultSocialCare.Api.V1.Domain.ClientDomains;
 using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Domain.ClientDomains;
 
-namespace LBH.AdultSocialCare.Api.V1.UseCase.Interfaces
+namespace LBH.AdultSocialCare.Api.V1.UseCase.ClientsUseCases.Interfaces
 {
     public interface IUpsertClientsUseCase
     {
